pyproject.toml
[project]
name = "crop-pest-detection"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "albumentations>=2.0.8",
    "dvc>=3.64.2",
    "dvc-s3>=3.2.2",
    "hydra-core>=1.3.2",
    "mlflow>=3.7.0",
    "omegaconf>=2.3.0",
    "onnx>=1.20.0",
    "onnxruntime>=1.23.2",
    "onnxscript>=0.5.7",
    "pycocotools>=2.0.11",
    "pydantic>=2.12.5",
    "pytorch-lightning>=2.6.0",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
    "tritonclient[http]>=2.64.0",
]

[dependency-groups]
dev = [
    "pre-commit>=4.5.1",
    "pytest>=9.0.2",
    "ruff>=0.14.10",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["crop_pest_detection*"]
